applications corrections

+++ routes/applications.js
@@ -3,9 +3,26 @@ const pool = require("../db_config");
 const auth = require("../middleware/auth");
 
 // Create new application
-router.post("/", (req, res) => {
-  res.send(`Here is placeholder code for adding a new job application_number`);
-});
+router.post("/", auth, (req, res) => {  
+  const account_id = res.locals.user.id
+    const insertQuery =
+      'INSERT INTO "applications" ("status", "date_added", "notes", "company", "position", "description", "salary","account_id") VALUES ($1, $2, $3, $4, $5, $6, $7,$8);';
+    const values = [
+      req.body.status,
+      req.body.date_added,
+      req.body.notes,
+      req.body.company,
+      req.body.position,
+      req.body.description,
+      req.body.salary,
+      account_id,
+    ];
+  
+    pool.query(insertQuery, values, (err, result) => {
+      if (err) return res.status(400).json({ error: err });
+      res.status(200).json({ messege : "application created"});
+    });
+  });
 
 //Get Count of Applications
 router.get("/count",auth, function (req, res) {
@@ -38,9 +55,10 @@ router.get("/:application_number", auth,  function (req, res, next) {
 });
 
 // Update an application
-router.put("/", auth, (req, res) => {
-  const text =
-    "UPDATE applications SET (status, date_added, notes, company, position, description, salary) = ($1, $2, $3, $4, $5, $6) WHERE id = $7;";
+router.put("/:application_id", auth, (req, res) => {
+  const { application_id } = req.params;    
+  const updateQuery =
+    'UPDATE "applications" SET ("status", "date_added", "notes", "company", "position", "description", "salary") = ($1, $2, $3, $4, $5, $6, $7) WHERE id = $8;';
   const values = [
     req.body.status,
     req.body.date_added,
@@ -49,6 +67,7 @@ router.put("/", auth, (req, res) => {
     req.body.position,
     req.body.description,
     req.body.salary,
+    application_id,
   ];
 
   pool.query(updateQuery, values, (err, results) => {
@@ -59,22 +78,21 @@ router.put("/", auth, (req, res) => {
 });
 
 // Get multiple applications
-router.get("/:applications_number", auth,  function (req, res, next) { 
-  let { applications_number } = req.params;                         
+router.get("/", auth,  function (req, res, next) { 
+  const account_id = res.locals.user.id                      
   const text =                                                 
-  "SELECT status, date_added, notes, company, position, description, salary FROM applications";
-  pool.query(text, [applications_number], (err, result) => {
+  "SELECT status, date_added, notes, company, position, description, salary FROM applications WHERE account_id = $1";
+  pool.query(text, [account_id], (err, result) => {
     if (err) {
       console.log(err);
       res.status(400).send(err);
     }
     res.status(200).send(result.rows);
-    console.log(result.affectedRows + " application(s) loaded");
   });
 });
 
 // Delete an application
-router.delete("/:application_number", auth,  (req, res) => {
+router.delete("/:applications_number", auth,  (req, res) => {
   let { applications_number } = req.params;
   let sql = "DELETE FROM applications WHERE id = $1";
   pool.query(sql, [applications_number], (err, result) => {
@@ -83,21 +101,19 @@ router.delete("/:application_number", auth,  (req, res) => {
       res.status(400).send(err);
     }
     res.status(204).send({"result": "Application successfully deleted"});
-    console.log(`Application ${application_number} deleted`);
   });
 });
 
 // Delete all applications
-router.delete("/:applications_number", auth,  (req, res) => {
-  let { applications_number } = req.params;
-  let sql =  "DELETE FROM applications";
-  pool.query(sql, [applications_number], (err, result) => {
+router.delete("/", auth,  (req, res) => {
+  const account_id = res.locals.user.id      
+  let sql =  "DELETE FROM applications WHERE account_id = $1";
+  pool.query(sql, [account_id], (err, result) => {
     if (err) {
       console.log(err);
       res.status(400).send(err);
     }
     res.status(204).send({"result": "Applications successfully deleted"});
-    console.log(result.affectedRows + " application(s) deleted");
   });
 });
